+++ projects/mk-kit/src/lib/components/directives/index.ts
@@ -2,3 +2,4 @@ export * from './click-outside';
 export * from './copy-to-clipboard';
 export * from './autosize';
 export * from './autofocus';
+export * from './scrollspy';